move cog to the core
