+++ martin-core/src/tiles/cog/model.rs
