+++ martin-core/src/tiles/cog/snapshots/martin__cog__image__tests__all_transparent.snap.png
