+++ martin-core/src/tiles/cog/snapshots/martin__cog__image__tests__down_padded_with_alpha.snap.png
