+++ martin-core/src/tiles/cog/snapshots/martin__cog__image__tests__not_padded.snap.png
