+++ martin-core/src/tiles/cog/snapshots/martin__cog__image__tests__right_padded.snap.png
