+++ martin/src/lib.rs
@@ -9,8 +9,6 @@ pub use source::TileSources;
 mod utils;
 pub use utils::{IdResolver, MartinError, MartinResult, NO_MAIN_CACHE};
 
-#[cfg(feature = "cog")]
-pub mod cog;
 #[cfg(feature = "postgres")]
 pub mod pg;
 #[cfg(feature = "pmtiles")]
